fix build error with AppleClang

+++ src/ainesmile/CodeEdit/findreplace.cpp
@@ -65,7 +65,7 @@ bool FindReplacer::findAllStringInFile(const QString &filePath, FindReplaceOptio
         while (pos >= 0)
         {
             // notify find result
-            int  from    = std::min(pos - 10, 0);
+            int  from    = std::min((int)pos - 10, 0);
             auto context = line.mid(from, fro.strToFind.length() + 20);
             notifier.addResult(filePath, context, lineNr, from, fro.strToFind.length());
 
@@ -107,7 +107,7 @@ bool FindReplacer::findAllRegexpInFile(const QString &filePath, FindReplaceOptio
         while (pos >= 0)
         {
             // notify find result
-            int  from    = std::min(pos - 10, 0);
+            int  from    = std::min((int)pos - 10, 0);
             auto context = line.mid(from, match.capturedLength() + 20);
             notifier.addResult(filePath, context, lineNr, from, match.capturedLength());
